requirements.txt
asgiref
Django
django-extensions
django-filter
djangorestframework
djangorestframework-jsonapi
inflection
python-dotenv
pytz
sqlparse
tzdata
gunicorn
whitenoise
django-cors-headers 
